+++ src/main/scala/com/dataintuitive/viash/platforms/NextFlowPlatform.scala
@@ -461,11 +461,14 @@ case class NextFlowPlatform(
         |    STUB=1 $$cli
         |    \"\"\"
         |  script:
+        |    def viash_temp = System.getenv("VIASH_TEMP") ?: "/tmp/"
         |    if (params.test)
         |      \"\"\"
         |      # Some useful stuff
         |      export NUMBA_CACHE_DIR=/tmp/numba-cache
         |      # Running the pre-hook when necessary
+        |      # Pass viash temp dir
+        |      export VIASH_TEMP="$${viash_temp}"
         |      # Adding NXF's `$$moduleDir` to the path in order to resolve our own wrappers
         |      export PATH="./:$${moduleDir}:\\$$PATH"
         |      ./$${params.$fname.tests.testScript} | tee $$output
@@ -475,6 +478,8 @@ case class NextFlowPlatform(
         |      # Some useful stuff
         |      export NUMBA_CACHE_DIR=/tmp/numba-cache
         |      # Running the pre-hook when necessary
+        |      # Pass viash temp dir
+        |      export VIASH_TEMP="$${viash_temp}"
         |      # Adding NXF's `$$moduleDir` to the path in order to resolve our own wrappers
         |      export PATH="$${moduleDir}:\\$$PATH"
         |      $$cli